Merging in changes from pr
This merges in some changes from this PR
https://github.com/electric-it/minimart/pull/15
Thank you to g3kk0 for creating that pr.
This branch and the pr branch do essesntialy the same thing but look at
it from differenct perspectives. As part of the requirements from GE the
load dependencies option should be off by default now and only loading
of the option is true.

Todd Pickell @tapickell <tapickell@gmail.com>